+++ eveuniverse/app_settings.py
@@ -1,5 +1,11 @@
 from .utils import clean_setting
 
+# Technical parameter defining the maximum number of objects processed per run
+# of Django batch methods, e.g. bulk_create and bulk_update
+EVEUNIVERSE_BULK_METHODS_BATCH_SIZE = clean_setting(
+    "EVEUNIVERSE_BULK_METHODS_BATCH_SIZE", 500
+)
+
 # when true will automatically load astroid belts with every solar system
 EVEUNIVERSE_LOAD_ASTEROID_BELTS = clean_setting(
     "EVEUNIVERSE_LOAD_ASTEROID_BELTS", False
@@ -28,3 +34,6 @@ EVEUNIVERSE_LOAD_STARS = clean_setting("EVEUNIVERSE_LOAD_STARS", False)
 
 # when true will automatically load stations be with every solar system
 EVEUNIVERSE_LOAD_STATIONS = clean_setting("EVEUNIVERSE_LOAD_STATIONS", False)
+
+# Global timeout for tasks in seconds to reduce task accumulation during outages
+EVEUNIVERSE_TASKS_TIME_LIMIT = clean_setting("EVEUNIVERSE_TASKS_TIME_LIMIT", 7200)